+++ src/javascripts/Data/clickEvents.js
@@ -1,0 +1,8 @@
+import pullForm from './data/components/pullForm';
+
+const clickEvents = () => {
+  document.querySelector("#infoForm").addEventListener("submit", pullForm);
+  document.querySelector("#bearsCard").addEventListener("click", cardRemoval);
+};
+
+export default clickEvents;
